Create new parameter set

Create a new parameter set within an rgeoprofile_project. The new parameter set becomes the active set once created.

new_set(project, spatial_prior = NULL, source_model = "uniform",
  name = "(no name)", sigma_model = "single",
  dispersal_model = "normal", sigma_prior_mean = 1,
  sigma_prior_sd = 1, expected_popsize_model = "single",
  expected_popsize_prior_mean = 1000, expected_popsize_prior_sd = 20,
  sentinel_radius = 0.2, n_binom = FALSE, alpha_prior_mean = 1,
  alpha_prior_sd = 100, dirichlet_scale = 1)

Arguments

project
an rgeoprofile_project, as produced by the function rgeoprofile_project().

spatial_prior
a raster file defining the spatial prior. Precision values are taken from this raster if it is defined.

source_model
choose prior type for source locations. Pick from "uniform" (default), "normal" (bivariate normal), "kernal" (KDE based on positive data) or "manual" (the current value of the raster)

name
an optional name for the parameter set.

sigma_model
set as "single" to assume the same dispersal distance for all sources, or "independent" to assume an independently drawn dispersal distance for each source.

dispersal_model
distribute points via a "normal", "cauchy" or "laplace" model

sigma_prior_mean
the prior mean of the parameter sigma (km).

sigma_prior_sd
the prior standard deviation of the parameter sigma (km). Set to 0 to use a fixed value for sigma (fixed at sigma_prior_mean).

expected_popsize_model
set as "single" to assume the same number of events for all sources, or "independent" to assume an independently drawn number of events for each source.

expected_popsize_prior_mean
the prior mean of the expected total population size.

expected_popsize_prior_sd
the prior standard deviation of the expected total population size. Set to 0 to use a fixed value (fixed at expected_popsize_prior_mean).

sentinel_radius
the observation radius of sentinel sites.

n_binom
set to true or false, decide if a negative binomial model should be run for a set of count data.

alpha_prior_mean
the prior mean alpha.

alpha_prior_sd
the prior standard deviation of alpha.

dirichlet_scale
a value to scale the concentration parameters for the dirichlet proposal on source weights