Add spatial prior to dynamic map

Add spatial prior to dynamic map

overlay_spatial_prior(myplot, project, col = col_hotcold(),
  opacity = 0.8, smoothing = 1)

Arguments

myplot
dynamic map produced by plot_map() function.

project
an RgeoProfile project, as produced by the function rgeoprofile_project().

col
set of plotting colours.

opacity
opacity of spatial prior.

smoothing
what level of smoothing to apply to spatial prior Smoothing is applied using the raster function disaggregate, with method = "bilinear".

Examples

library(silverblaze)
p <- rgeoprofile_file("tutorial1_project.rds")
plot1 <- plot_map()
plot1 <- overlay_spatial_prior(myplot = plot1, project = p)
plot1