Draw from spherical normal distribution

Draw from normal distribution converted to spherical coordinate system. Points are first drawn from an ordinary cartesian 2D normal distribution. The distances to points are then assumed to be great circle distances, and are combined with a random bearing from the point centre_lat, centre_lon to produce a final set of lat/lon points.

rnorm_sphere(n, centre_lon, centre_lat, sigma = 1)

Arguments

n
The number of points to draw

centre_lon
The mean longitude of the normal distribution

centre_lat
The mean latitude of the normal distribution

sigma
The standard deviation of the normal distribution

Examples

rnorm_sphere(n = 100, centre_lat = 0, centre_lon = 0, sigma = 1)

#> $longitude
#>   [1]  0.0042102140  0.0032640991 -0.0117320421  0.0066349819  0.0169837329
#>   [6] -0.0008763449 -0.0084162775 -0.0001434446 -0.0074354917 -0.0136013368
#>  [11]  0.0084119234  0.0015872002  0.0021915160  0.0146009261  0.0010075827
#>  [16] -0.0012050641 -0.0171778294 -0.0025112408 -0.0028188874  0.0095985304
#>  [21]  0.0006298385 -0.0057477742 -0.0004493451 -0.0022616450  0.0040001566
#>  [26]  0.0247800659  0.0004184668  0.0051954625  0.0010629521 -0.0171925158
#>  [31]  0.0077529300 -0.0021874806 -0.0018533867  0.0001724682  0.0002658463
#>  [36]  0.0049447179 -0.0204516063  0.0241248164 -0.0032485408  0.0019187544
#>  [41]  0.0096618249 -0.0059812824  0.0100180148 -0.0022113996 -0.0105900815
#>  [46] -0.0087760355  0.0095782907  0.0011841425  0.0043943445 -0.0152835262
#>  [51] -0.0132266494  0.0025554255  0.0120268116  0.0021286608  0.0118556973
#>  [56]  0.0047116340  0.0054566163 -0.0009886753  0.0015484674 -0.0008123328
#>  [61]  0.0173060356  0.0116767267  0.0067340418  0.0050022457 -0.0049305960
#>  [66]  0.0099872806 -0.0234932871 -0.0014001878  0.0039020647 -0.0034349689
#>  [71]  0.0038148105  0.0095607060  0.0094312992 -0.0003426676  0.0043720424
#>  [76]  0.0150445948 -0.0031868465  0.0085107111  0.0118425041 -0.0026677478
#>  [81] -0.0034822953 -0.0070635657 -0.0095034630 -0.0071544760 -0.0157945652
#>  [86] -0.0062101565 -0.0050230889 -0.0048263293  0.0020426034  0.0087994570
#>  [91] -0.0018785268 -0.0125852006  0.0023250817 -0.0039731980  0.0051135414
#>  [96]  0.0191272257  0.0038208438 -0.0151471077  0.0022429242  0.0096482663
#> 
#> $latitude
#>   [1]  0.0183404884  0.0040420349  0.0125168843  0.0038362051  0.0009675261
#>   [6]  0.0002005014  0.0054284042 -0.0023620733 -0.0047507930  0.0017280413
#>  [11] -0.0103080212  0.0076099215  0.0007349223 -0.0117371991 -0.0084977983
#>  [16]  0.0040859921 -0.0076910207 -0.0025801107  0.0080485833  0.0006052834
#>  [21] -0.0014629834 -0.0074401791  0.0168758205  0.0068927623  0.0088129623
#>  [26]  0.0118870617 -0.0100698014  0.0046279073 -0.0135716609  0.0137842751
#>  [31]  0.0038594150  0.0010981024 -0.0102343914 -0.0050183506  0.0094657065
#>  [36]  0.0060945554  0.0003462347 -0.0032050129  0.0070402861  0.0072342475
#>  [41] -0.0170876574  0.0084157103 -0.0027793669  0.0023658155 -0.0161031793
#>  [46] -0.0070889821 -0.0101895086  0.0032704061 -0.0025710518  0.0046555103
#>  [51] -0.0009254799 -0.0087600183  0.0114274305  0.0086412646  0.0069132773
#>  [56]  0.0093163492 -0.0042617688 -0.0114693621 -0.0027485128  0.0198909209
#>  [61] -0.0093679487 -0.0103109366 -0.0150665804  0.0137230960  0.0049839101
#>  [66]  0.0179244712 -0.0013860411  0.0230622782  0.0095507878  0.0102765019
#>  [71]  0.0101069255 -0.0035703202 -0.0074037654 -0.0052060345  0.0158621389
#>  [76]  0.0011960271  0.0033859398  0.0102406439  0.0111629469  0.0055046661
#>  [81] -0.0038615079  0.0122349226 -0.0006372362 -0.0024475369 -0.0220035219
#>  [86]  0.0005889355 -0.0098791279 -0.0056943081 -0.0185588903  0.0238224180
#>  [91] -0.0103727609 -0.0030634300  0.0070719285 -0.0114259989  0.0048755961
#>  [96]  0.0006754436  0.0050228409  0.0037358395 -0.0130608456  0.0084644700
#>