Bind data to project

Load data into a rgeoprofile_project prior to analysis. Data must be formatted as a dataframe with samples in rows and loci in columns. If individuals are polyploid then multiple rows can be used per sample. Ploidy is allowed to vary between samples, and can be specified in multiple ways.

bind_data(project, df, name = NULL, check_delete_output = TRUE)

Arguments

project
an rgeoprofile_project, as produced by the function rgeoprofile_project()

df
a dataframe with three columns named "longitude", "latitude", "counts" (in that order). Counts must be positive integers or zero

name
optional name of the data set to aid in record keeping

check_delete_output
whether to prompt the user before overwriting existing data