Calculate pairwise great circle distance between points

Analogue of the dist() function, but calculating great circle distances. Points should be input as a two-column matrix or dataframe with longitude in the first column and latitude in the second.

dist_gc(x)

Arguments

x
a two-column matrix or dataframe with longitude in the first column and latitude in the second

Examples

london_lon <- runif(20, min = -0.2, max = 0)
london_lat <- runif(20, min = 51.47, max = 51.53)
some_data  <- data.frame(longitude = london_lon, latitude = london_lat)
head(some_data)

#>     longitude latitude
#> 1 -0.18384997 51.48739
#> 2 -0.03313339 51.51070
#> 3 -0.07984782 51.51412
#> 4 -0.16855831 51.48176
#> 5 -0.19852011 51.52883
#> 6 -0.10672130 51.51449

distance_matrix <- dist_gc(some_data)
head(distance_matrix)

#>           [,1]      [,2]     [,3]     [,4]      [,5]     [,6]     [,7]     [,8]
#> [1,]  0.000000 10.750022 7.788216 1.230421  4.718320 6.130372 5.990944 3.308461
#> [2,] 10.750022  0.000000 3.254958 9.912062 11.619390 5.109775 6.261946 7.602526
#> [3,]  7.788216  3.254958 0.000000 7.117520  8.372088 1.860071 4.780219 4.516941
#> [4,]  1.230421  9.912062 7.117520 0.000000  5.630304 5.618867 4.814723 2.888278
#> [5,]  4.718320 11.619390 8.372088 5.630304  0.000000 6.548465 9.192897 4.929269
#> [6,]  6.130372  5.109775 1.860071 5.618867  6.548465 0.000000 4.624433 2.822148
#>           [,9]      [,10]     [,11]     [,12]      [,13]    [,14]    [,15]
#> [1,]  9.425713  9.9395591 11.129654 1.3731921  0.6481672 4.043951 4.627977
#> [2,]  1.408966  0.8582072  4.353442 9.6161488 11.3594885 7.119154 6.126218
#> [3,]  1.847466  2.3975927  6.036787 6.8059058  8.3613291 3.936235 3.235487
#> [4,]  8.639836  9.1310421  9.997286 0.3151568  1.8410705 3.706989 3.884212
#> [5,] 10.217302 10.7654695 13.573713 5.5452284  4.5500049 4.901971 6.408599
#> [6,]  3.700809  4.2515756  7.375834 5.3038386  6.6660230 2.138940 1.963023
#>         [,16]    [,17]      [,18]    [,19]     [,20]
#> [1,] 1.852873 5.262366  4.4009024 4.326970 12.401247
#> [2,] 8.992371 5.965391 10.8161222 7.007785  3.647174
#> [3,] 5.961092 2.736371  7.5672648 4.725403  6.235863
#> [4,] 1.663514 4.819767  5.1855426 3.207817 11.334404
#> [5,] 4.491633 5.778699  0.8071014 7.507415 14.337152
#> [6,] 4.278243 0.900866  5.7415853 3.895488  7.870122